// packages/cactus-plugin-ledger-connector-daml/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface CreateIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequest
     */
    'templateId'?: string;
    /**
     * 
     * @type {CreateIOUEndpointRequestPayload}
     * @memberof CreateIOUEndpointRequest
     */
    'payload'?: CreateIOUEndpointRequestPayload;
}
/**
 * 
 * @export
 * @interface CreateIOUEndpointRequestPayload
 */
export interface CreateIOUEndpointRequestPayload {
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'issuer'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'owner'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'currency'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'amount'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'observers'?: Array<string>;
}
/**
 * This method is the response for creating a simple IOU contract.
 * @export
 * @interface CreateIOUEndpointResponse
 */
export interface CreateIOUEndpointResponse {
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointResponse
     */
    'agreementText'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointResponse
     */
    'completionOffset'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointResponse
     */
    'contractId'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateIOUEndpointResponse
     */
    'observers'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointResponse
     */
    'templateId'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateIOUEndpointResponse
     */
    'signatories'?: Array<string>;
    /**
     * 
     * @type {CreateIOUEndpointRequestPayload}
     * @memberof CreateIOUEndpointResponse
     */
    'payload'?: CreateIOUEndpointRequestPayload;
}
/**
 * This method is the request for creating a simple IOU contract.
 * @export
 * @interface ExerciseIOUEndpointRequest
 */
export interface ExerciseIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'previousParticipantToken'?: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'templateId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'contractId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'choice'?: string;
    /**
     * 
     * @type {ExerciseIOUEndpointRequestArgument}
     * @memberof ExerciseIOUEndpointRequest
     */
    'argument'?: ExerciseIOUEndpointRequestArgument;
}
/**
 * 
 * @export
 * @interface ExerciseIOUEndpointRequestArgument
 */
export interface ExerciseIOUEndpointRequestArgument {
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequestArgument
     */
    'newOwner'?: string;
}
// ...
export interface QueryIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof QueryIOUEndpointRequest
     */
    'participantToken'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof QueryIOUEndpointRequest
     */
    'templateIds'?: Array<string>;
    /**
     * 
     * @type {QueryIOUEndpointRequestQuery}
     * @memberof QueryIOUEndpointRequest
     */
    'query'?: QueryIOUEndpointRequestQuery;
    /**
     * 
     * @type {Array<string>}
     * @memberof QueryIOUEndpointRequest
     */
    'readers'?: Array<string>;
}
/**
 * 
 * @export
 * @interface QueryIOUEndpointRequestQuery
 */
export interface QueryIOUEndpointRequestQuery {
    /**
     * 
     * @type {number}
     * @memberof QueryIOUEndpointRequestQuery
     */
    'amount'?: number;
}
/**
 * This method is the response for querying a simple IOU contract.
 * @export
 * @interface QueryIOUEndpointResponse
 */
export interface QueryIOUEndpointResponse {
    /**
     * 
     * @type {string}
     * @memberof QueryIOUEndpointResponse
     */
    'completionOffset'?: string;
    /**
     * 
     * @type {Array<QueryIOUEndpointResponseEventsInner>}
     * @memberof QueryIOUEndpointResponse
     */
    'events'?: Array<QueryIOUEndpointResponseEventsInner>;
}
/**
 * 
 * @export
 * @interface QueryIOUEndpointResponseEventsInner
 */
export interface QueryIOUEndpointResponseEventsInner {
    /**
     * 
     * @type {QueryIOUEndpointResponseEventsInnerArchived}
     * @memberof QueryIOUEndpointResponseEventsInner
     */
    'archived'?: QueryIOUEndpointResponseEventsInnerArchived;
}
/**
 * 
 * @export
 * @interface QueryIOUEndpointResponseEventsInnerArchived
 */
export interface QueryIOUEndpointResponseEventsInnerArchived {
    /**
     * 
     * @type {string}
     * @memberof QueryIOUEndpointResponseEventsInnerArchived
     */
    'contractId'?: string;
    /**
     * 
     * @type {string}
     * @memberof QueryIOUEndpointResponseEventsInnerArchived
     */
    'templateId'?: string;
}
/**
 * This method is the request for querying a simple IOU contract.
 * @export
 * @interface QueryRawIOUEndpointRequest
 */
export interface QueryRawIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof QueryRawIOUEndpointRequest
     */
    'participantToken'?: string;
}